Extract basic logic of safety to SafetyFragment
