+++ app/src/main/kotlin/com/mapbox/vision/teaser/MainActivity.kt
@@ -1,6 +1,5 @@
 package com.mapbox.vision.teaser
 
-import android.animation.Animator
 import android.annotation.SuppressLint
 import android.app.Activity
 import android.content.Intent
@@ -9,7 +8,6 @@ import android.os.Environment
 import android.text.method.LinkMovementMethod
 import android.view.ViewGroup
 import android.view.WindowManager
-import android.view.animation.AccelerateDecelerateInterpolator
 import android.widget.ImageView
 import android.widget.SeekBar
 import android.widget.TextView
@@ -19,15 +17,11 @@ import androidx.appcompat.app.AppCompatActivity
 import androidx.core.text.HtmlCompat
 import androidx.core.view.setPadding
 import androidx.fragment.app.Fragment
-import com.mapbox.services.android.navigation.v5.navigation.NavigationConstants
-import com.mapbox.services.android.navigation.v5.utils.DistanceFormatter
-import com.mapbox.services.android.navigation.v5.utils.LocaleUtils
 import com.mapbox.vision.VisionManager
 import com.mapbox.vision.VisionReplayManager
 import com.mapbox.vision.mobile.core.interfaces.VisionEventsListener
 import com.mapbox.vision.mobile.core.models.Country
 import com.mapbox.vision.mobile.core.models.classification.FrameSignClassifications
-import com.mapbox.vision.mobile.core.models.detection.DetectionClass
 import com.mapbox.vision.mobile.core.models.position.VehicleState
 import com.mapbox.vision.mobile.core.models.road.LaneDirection
 import com.mapbox.vision.mobile.core.models.road.LaneEdgeType
@@ -36,11 +30,6 @@ import com.mapbox.vision.mobile.core.utils.SystemInfoUtils
 import com.mapbox.vision.performance.ModelPerformance
 import com.mapbox.vision.performance.ModelPerformanceMode
 import com.mapbox.vision.performance.ModelPerformanceRate
-import com.mapbox.vision.safety.VisionSafetyManager
-import com.mapbox.vision.safety.core.VisionSafetyListener
-import com.mapbox.vision.safety.core.models.CollisionDangerLevel
-import com.mapbox.vision.safety.core.models.CollisionObject
-import com.mapbox.vision.safety.core.models.RoadRestrictions
 import com.mapbox.vision.teaser.MainActivity.VisionMode.Camera
 import com.mapbox.vision.teaser.MainActivity.VisionMode.Replay
 import com.mapbox.vision.teaser.ar.ArMapActivity
@@ -49,8 +38,7 @@ import com.mapbox.vision.teaser.models.UiSign
 import com.mapbox.vision.teaser.recorder.RecorderFragment
 import com.mapbox.vision.teaser.replayer.ArReplayNavigationActivity
 import com.mapbox.vision.teaser.replayer.ReplayModeFragment
-import com.mapbox.vision.teaser.utils.PermissionsUtils
-import com.mapbox.vision.teaser.utils.SoundsPlayer
+import com.mapbox.vision.teaser.utils.allPermissionsGrantedByRequest
 import com.mapbox.vision.teaser.utils.classification.SignResources
 import com.mapbox.vision.teaser.utils.classification.Tracker
 import com.mapbox.vision.teaser.utils.dpToPx
@@ -74,35 +62,31 @@ class MainActivity : AppCompatActivity(), ReplayModeFragment.OnSelectModeItemLis
         Segmentation,
         Classification,
         Detection,
-        Safety,
         Lanes
     }
 
     companion object {
         private val BASE_SESSION_PATH = "${Environment.getExternalStorageDirectory().absolutePath}/MapboxVisionTelemetry"
         private const val TRACKER_DEFAULT_COUNT = 5
-        private const val CALIBRATION_READY_VALUE = 1f
         private const val START_AR_MAP_ACTIVITY_FOR_NAVIGATION_RESULT_CODE = 100
         private const val START_AR_MAP_ACTIVITY_FOR_RECORDING_RESULT_CODE = 110
     }
 
-    private var visionManagerMode = Camera
+    var visionMode = Camera
     private var sessionPath = ""
 
-    private val signResources: SignResources = SignResources.Impl(this)
+    val signResources: SignResources = SignResources.Impl(this)
 
     private var signSize = 0
     private var margin = 0
 
     private var lineHeight = 0
     private var tracker: Tracker<UiSign> = Tracker(TRACKER_DEFAULT_COUNT)
-    private var visionMode: VisionFeature = VisionFeature.Detection
+    private var visionFeature: VisionFeature = VisionFeature.Detection
     private var isPermissionsGranted = false
     private var visionManagerWasInit = false
-    private lateinit var soundsPlayer: SoundsPlayer
     private var country = Country.Unknown
     private var modelPerformance = ModelPerformance.On(ModelPerformanceMode.FIXED, ModelPerformanceRate.HIGH)
-    private var lastSpeed = 0f
     private var calibrationProgress = 0f
     private var isProgressChanging = false
 
@@ -110,10 +94,11 @@ class MainActivity : AppCompatActivity(), ReplayModeFragment.OnSelectModeItemLis
 
         override fun onCountryUpdated(country: Country) {
             this@MainActivity.country = country
+            requireSafetyFragment()?.country = country
         }
 
         override fun onFrameSignClassificationsUpdated(frameSignClassifications: FrameSignClassifications) {
-            if (visionMode == VisionFeature.Classification) {
+            if (visionFeature == VisionFeature.Classification) {
                 runOnUiThreadIfPossible {
                     tracker.update(UiSign.getUiSigns(frameSignClassifications))
                     drawSigns(tracker.getCurrent())
@@ -122,7 +107,7 @@ class MainActivity : AppCompatActivity(), ReplayModeFragment.OnSelectModeItemLis
         }
 
         override fun onRoadDescriptionUpdated(roadDescription: RoadDescription) {
-            if (visionMode == VisionFeature.Lanes) {
+            if (visionFeature == VisionFeature.Lanes) {
                 runOnUiThreadIfPossible {
                     drawLanesDetection(roadDescription)
                 }
@@ -130,11 +115,11 @@ class MainActivity : AppCompatActivity(), ReplayModeFragment.OnSelectModeItemLis
         }
 
         override fun onVehicleStateUpdated(vehicleState: VehicleState) {
-            lastSpeed = vehicleState.speed
+            requireSafetyFragment()?.lastSpeed = vehicleState.speed
         }
 
         override fun onCameraUpdated(camera: com.mapbox.vision.mobile.core.models.Camera) {
-            calibrationProgress = camera.calibrationProgress
+            requireSafetyFragment()?.calibrationProgress = camera.calibrationProgress
             runOnUiThreadIfPossible {
                 fps_performance_view.setCalibrationProgress(calibrationProgress)
             }
@@ -144,12 +129,12 @@ class MainActivity : AppCompatActivity(), ReplayModeFragment.OnSelectModeItemLis
         override fun onUpdateCompleted() {
             runOnUiThread {
                 if (visionManagerWasInit) {
-                    val frameStatistics = when (visionManagerMode) {
+                    val frameStatistics = when (visionMode) {
                         Camera -> VisionManager.getFrameStatistics()
                         Replay -> VisionReplayManager.getFrameStatistics()
                     }
                     fps_performance_view.showInfo(frameStatistics)
-                    if (visionManagerMode == Replay && !isProgressChanging) {
+                    if (visionMode == Replay && !isProgressChanging) {
                         playback_seek_bar_view.setProgress(VisionReplayManager.getProgress())
                     }
                 }
@@ -157,159 +142,12 @@ class MainActivity : AppCompatActivity(), ReplayModeFragment.OnSelectModeItemLis
         }
     }
 
-    private fun runOnUiThreadIfPossible(action: () -> Unit) {
+    fun runOnUiThreadIfPossible(action: () -> Unit) {
         runOnUiThread {
-            if (visionManagerMode == Replay && isProgressChanging) {
+            if (visionMode == Replay && isProgressChanging) {
                 return@runOnUiThread
             }
             action.invoke()
-        }
-    }
-
-    private val visionSafetyListener = object : VisionSafetyListener {
-
-        private var currentDangerLevel: CollisionDangerLevel = CollisionDangerLevel.None
-
-        private val distanceFormatter by lazy {
-            LocaleUtils().let { localeUtils ->
-                val language = localeUtils.inferDeviceLanguage(this@MainActivity)
-                val unitType = localeUtils.getUnitTypeForDeviceLocale(this@MainActivity)
-                val roundingIncrement = NavigationConstants.ROUNDING_INCREMENT_FIVE
-                DistanceFormatter(this@MainActivity, language, unitType, roundingIncrement)
-            }
-        }
-
-        override fun onCollisionsUpdated(collisions: Array<CollisionObject>) {
-
-            fun updateCollisionDangerLevelIfChanged(collision: CollisionObject) {
-                if (currentDangerLevel != collision.dangerLevel) {
-                    soundsPlayer.stop()
-                    when (collision.dangerLevel) {
-                        CollisionDangerLevel.None -> Unit
-                        CollisionDangerLevel.Warning -> soundsPlayer.playWarning()
-                        CollisionDangerLevel.Critical -> soundsPlayer.playCritical()
-                    }
-                    currentDangerLevel = collision.dangerLevel
-                }
-            }
-
-            fun showCollision(collision: CollisionObject) {
-                distance_to_car_label.show()
-                safety_mode.show()
-                distance_to_car_label.text =
-                        distanceFormatter.formatDistance(collision.`object`.position.y)
-
-                when (currentDangerLevel) {
-                    CollisionDangerLevel.None -> safety_mode.clean()
-                    CollisionDangerLevel.Warning -> safety_mode.drawWarnings(collisions)
-                    CollisionDangerLevel.Critical -> safety_mode.drawCritical()
-                }
-            }
-
-            fun hideCollision() {
-                soundsPlayer.stop()
-                currentDangerLevel = CollisionDangerLevel.None
-                distance_to_car_label.hide()
-                safety_mode.hide()
-            }
-
-            fun calibrationReady() {
-                distance_to_car_label.show()
-                safety_mode.show()
-                calibration_progress.hide()
-
-                val collision = collisions.firstOrNull { it.`object`.objectClass == DetectionClass.Car }
-                if (collision == null) {
-                    hideCollision()
-                } else {
-                    updateCollisionDangerLevelIfChanged(collision)
-                    showCollision(collision)
-                }
-            }
-
-            fun calibrationInProgress() {
-                distance_to_car_label.hide()
-                safety_mode.hide()
-                calibration_progress.show()
-                val progress = (calibrationProgress * 100).toInt()
-                calibration_progress.text = getString(R.string.calibration_progress, progress)
-            }
-
-            if (visionMode == VisionFeature.Safety) {
-                runOnUiThreadIfPossible {
-                    if (calibrationProgress == CALIBRATION_READY_VALUE) {
-                        calibrationReady()
-                    } else {
-                        calibrationInProgress()
-                    }
-                }
-            }
-        }
-
-        val speedLimitTranslation by lazy {
-            resources.getDimension(R.dimen.speed_limit_translation)
-        }
-
-        override fun onRoadRestrictionsUpdated(roadRestrictions: RoadRestrictions) {
-            runOnUiThreadIfPossible {
-                val imageResource = signResources.getSpeedSignResource(
-                        UiSign.WithNumber(
-                                signType = UiSign.SignType.SpeedLimit,
-                                signNumber = UiSign.SignNumber.fromNumber(roadRestrictions.speedLimits.car.max)
-                        ),
-                        speed = lastSpeed,
-                        country = country
-                )
-
-                speed_limit_current.animate().cancel()
-                speed_limit_next.animate().cancel()
-
-                speed_limit_current.apply {
-                    show()
-                    translationY = 0f
-                    alpha = 1f
-                    animate()
-                            .translationY(speedLimitTranslation / 2)
-                            .alpha(0f)
-                            .scaleX(0.5f)
-                            .scaleY(0.5f)
-                            .setDuration(500L)
-                            .setListener(
-                                    object : Animator.AnimatorListener {
-                                        override fun onAnimationRepeat(animation: Animator?) {}
-
-                                        override fun onAnimationEnd(animation: Animator?) {
-                                            setImageResource(imageResource)
-                                            translationY = 0f
-                                            alpha = 1f
-                                            scaleX = 1f
-                                            scaleY = 1f
-                                            speed_limit_next.hide()
-                                        }
-
-                                        override fun onAnimationCancel(animation: Animator?) {}
-
-                                        override fun onAnimationStart(animation: Animator?) {}
-                                    }
-                            )
-                            .setInterpolator(AccelerateDecelerateInterpolator())
-                            .start()
-                }
-
-                if (roadRestrictions.speedLimits.car.max != 0f) {
-                    speed_limit_next.apply {
-                        translationY = -speedLimitTranslation
-                        setImageResource(imageResource)
-                        show()
-                        animate().translationY(0f)
-                                .setDuration(500L)
-                                .setInterpolator(AccelerateDecelerateInterpolator())
-                                .start()
-                    }
-                } else {
-                    speed_limit_next.hide()
-                }
-            }
         }
     }
 
@@ -345,8 +183,6 @@ class MainActivity : AppCompatActivity(), ReplayModeFragment.OnSelectModeItemLis
         if (!PermissionsUtils.requestPermissionsIfNotGranted(this)) {
             onPermissionsGranted()
         }
-
-        soundsPlayer = SoundsPlayer(this)
     }
 
     private fun createSessionFolderIfNotExist() {
@@ -370,8 +206,10 @@ class MainActivity : AppCompatActivity(), ReplayModeFragment.OnSelectModeItemLis
         segm_container.setOnClickListener { setVisionMode(VisionFeature.Segmentation) }
         sign_detection_container.setOnClickListener { setVisionMode(VisionFeature.Classification) }
         det_container.setOnClickListener { setVisionMode(VisionFeature.Detection) }
-        distance_container.setOnClickListener { setVisionMode(VisionFeature.Safety) }
-        safety_mode_container.hide()
+        distance_container.setOnClickListener {
+            vision_view.visualizationMode = VisualizationMode.Clear
+            showSafetyFragment()
+        }
         line_detection_container.setOnClickListener { setVisionMode(VisionFeature.Lanes) }
 
         initRootLongTap()
@@ -392,7 +230,7 @@ class MainActivity : AppCompatActivity(), ReplayModeFragment.OnSelectModeItemLis
 
     private fun initRootTap() {
         root.setOnClickListener {
-            if (visionManagerMode == Replay) {
+            if (visionMode == Replay) {
                 playback_seek_bar_view.toggleVisibleGone()
             }
         }
@@ -400,7 +238,7 @@ class MainActivity : AppCompatActivity(), ReplayModeFragment.OnSelectModeItemLis
 
     private fun initArNavigationButton() {
         ar_navigation_button_container.setOnClickListener {
-            when (visionManagerMode) {
+            when (visionMode) {
                 Camera -> startArMapActivityForNavigation()
                 Replay -> startArSession()
             }
@@ -415,7 +253,7 @@ class MainActivity : AppCompatActivity(), ReplayModeFragment.OnSelectModeItemLis
 
     private fun tryToInitVisionManager() {
         if (isPermissionsGranted && !visionManagerWasInit) {
-            visionManagerWasInit = when (visionManagerMode) {
+            visionManagerWasInit = when (visionMode) {
                 Camera -> initVisionManagerCamera(vision_view)
                 Replay -> initVisionManagerReplay(vision_view, sessionPath)
             }
@@ -424,7 +262,7 @@ class MainActivity : AppCompatActivity(), ReplayModeFragment.OnSelectModeItemLis
 
     private fun stopVisionManager() {
         if (isPermissionsGranted && visionManagerWasInit) {
-            when (visionManagerMode) {
+            when (visionMode) {
                 Camera -> destroyVisionManagerCamera()
                 Replay -> {
                     destroyVisionManagerReplay()
@@ -445,15 +283,12 @@ class MainActivity : AppCompatActivity(), ReplayModeFragment.OnSelectModeItemLis
         super.onPause()
         stopVisionManager()
         vision_view.onPause()
-        soundsPlayer.stop()
     }
 
     private fun onBackClick() {
-        soundsPlayer.stop()
         dashboard_container.show()
         hideLineDetectionContainer()
         hideSignsContainer()
-        safety_mode_container.hide()
         back.hide()
         playback_seek_bar_view.hide()
     }
@@ -543,17 +378,15 @@ class MainActivity : AppCompatActivity(), ReplayModeFragment.OnSelectModeItemLis
 
     private fun setVisionMode(mode: VisionFeature) {
         fps_performance_view.resetAverageFps()
-        soundsPlayer.stop()
 
         back.show()
         dashboard_container.hide()
-        safety_mode_container.hide()
         hideLineDetectionContainer()
         hideSignsContainer()
         playback_seek_bar_view.hide()
 
-        visionMode = mode
-        when (visionMode) {
+        visionFeature = mode
+        when (visionFeature) {
             VisionFeature.Classification -> {
                 vision_view.visualizationMode = VisualizationMode.Clear
                 tracker = Tracker(TRACKER_DEFAULT_COUNT)
@@ -564,13 +397,6 @@ class MainActivity : AppCompatActivity(), ReplayModeFragment.OnSelectModeItemLis
             }
             VisionFeature.Segmentation -> {
                 vision_view.visualizationMode = VisualizationMode.Segmentation
-            }
-            VisionFeature.Safety -> {
-                vision_view.visualizationMode = VisualizationMode.Clear
-                safety_mode_container.show()
-                safety_mode.hide()
-                calibration_progress.hide()
-                distance_to_car_label.hide()
             }
             VisionFeature.Lanes -> {
                 vision_view.visualizationMode = VisualizationMode.Clear
@@ -585,14 +411,10 @@ class MainActivity : AppCompatActivity(), ReplayModeFragment.OnSelectModeItemLis
         VisionManager.visionEventsListener = visionEventsListener
         VisionManager.start()
         VisionManager.setModelPerformance(modelPerformance)
-
-        VisionSafetyManager.create(VisionManager)
-        VisionSafetyManager.visionSafetyListener = visionSafetyListener
         return true
     }
 
     private fun destroyVisionManagerCamera() {
-        VisionSafetyManager.destroy()
         VisionManager.stop()
         VisionManager.destroy()
     }
@@ -608,9 +430,6 @@ class MainActivity : AppCompatActivity(), ReplayModeFragment.OnSelectModeItemLis
         VisionReplayManager.start()
         VisionReplayManager.setModelPerformance(modelPerformance)
         visionView.setVisionManager(VisionReplayManager)
-
-        VisionSafetyManager.create(VisionReplayManager)
-        VisionSafetyManager.visionSafetyListener = visionSafetyListener
 
         playback_seek_bar_view.setDuration(VisionReplayManager.getDuration())
         playback_seek_bar_view.onSeekBarChangeListener = object : SeekBar.OnSeekBarChangeListener {
@@ -634,7 +453,6 @@ class MainActivity : AppCompatActivity(), ReplayModeFragment.OnSelectModeItemLis
     }
 
     private fun destroyVisionManagerReplay() {
-        VisionSafetyManager.destroy()
         VisionReplayManager.stop()
         VisionReplayManager.destroy()
     }
@@ -647,6 +465,11 @@ class MainActivity : AppCompatActivity(), ReplayModeFragment.OnSelectModeItemLis
     private fun showRecorderFragment(jsonRoute: String?, stateLoss: Boolean = false) {
         val fragment = RecorderFragment.newInstance(BASE_SESSION_PATH, jsonRoute)
         showFragment(fragment, RecorderFragment.TAG, stateLoss)
+    }
+
+    private fun showSafetyFragment(stateLoss: Boolean = false) {
+        val fragment = SafetyFragment.newInstance()
+        showFragment(fragment, SafetyFragment.TAG, stateLoss)
     }
 
     private fun showFragment(fragment: Fragment, tag: String, stateLoss: Boolean = false) {
@@ -679,6 +502,7 @@ class MainActivity : AppCompatActivity(), ReplayModeFragment.OnSelectModeItemLis
                 if (supportFragmentManager.popBackStackImmediate() && supportFragmentManager.backStackEntryCount == 0) {
                     showDashboardView()
                     title_teaser.show()
+                    playback_seek_bar_view.hide()
                 }
             }
         } else {
@@ -688,14 +512,14 @@ class MainActivity : AppCompatActivity(), ReplayModeFragment.OnSelectModeItemLis
 
     override fun onSessionSelected(sessionName: String) {
         stopVisionManager()
-        visionManagerMode = Replay
+        visionMode = Replay
         sessionPath = "$BASE_SESSION_PATH/$sessionName"
         tryToInitVisionManager()
     }
 
     override fun onCameraSelected() {
         stopVisionManager()
-        visionManagerMode = Camera
+        visionMode = Camera
         sessionPath = ""
         tryToInitVisionManager()
     }
@@ -757,4 +581,8 @@ class MainActivity : AppCompatActivity(), ReplayModeFragment.OnSelectModeItemLis
             else -> super.onActivityResult(requestCode, resultCode, data)
         }
     }
+
+    private fun requireSafetyFragment() = supportFragmentManager.findFragmentById(R.id.fragment_container) as? SafetyFragment
+
+    fun isCameraMode() = visionMode == Camera
 }
